Add backward-compatibility for no pos embeddings
We call the type "nope" for forward-compatibility since upstream
maintainers seem to want to use that name according to their comment
removed in this commit.

This automatically chooses the "nope" position embedding when the
deprecated `--no-position-embedding` argument is given together with
`--position-embedding-type=learned_absolute` (the default). Upstream
throws an error here, we instead print a warning and change the position
embedding type to "nope" as described above.